+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+### Added
+- PS Provider base classes
+- Sandbox project
+
 ## 0.1.11
 ### Added
 - Evironment variables management when starting application